requirements.txt
flask>=3.0.0
groq>=0.9.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-dotenv>=1.0.0
httpx>=0.27.0
certifi>=2024.2.2
urllib3>=2.0.0
